reorg pretrain datasets
