+++ s3prl/pretrain/mockingjay/dataset.py
@@ -11,93 +11,20 @@
 # IMPORTATION #
 ###############
 import os
-import random
-#-------------#
 import numpy as np
-import pandas as pd
 #-------------#
 import torch
 from torch.nn.utils.rnn import pad_sequence
-from torch.utils.data.dataset import Dataset
 import torchaudio
 #-------------#
 from pretrain.mockingjay.task import generate_masked_acoustic_model_data
+from pretrain.bucket_dataset import FeatDataset
 
 
 HALF_BATCHSIZE_TIME = 99999
 
 
-####################
-# ACOUSTIC DATASET #
-####################
-class AcousticDataset(Dataset):
-    
-    def __init__(self, extracter, task_config, bucket_size, file_path, sets, 
-                 max_timestep=0, libri_root=None, **kwargs):
-        super(AcousticDataset, self).__init__()
-
-        self.extracter = extracter
-        self.task_config = task_config
-        self.libri_root = libri_root
-        self.sample_length = task_config['sequence_length'] 
-        if self.sample_length > 0:
-            print('[Dataset] - Sampling random segments for training, sample length:', self.sample_length)
-        
-        # Read file
-        self.root = file_path
-        tables = [pd.read_csv(os.path.join(file_path, s + '.csv')) for s in sets]
-        self.table = pd.concat(tables, ignore_index=True).sort_values(by=['length'], ascending=False)
-        print('[Dataset] - Training data from these sets:', str(sets))
-
-        # Drop seqs that are too long
-        if max_timestep > 0:
-            self.table = self.table[self.table.length < max_timestep]
-        # Drop seqs that are too short
-        if max_timestep < 0:
-            self.table = self.table[self.table.length > (-1 * max_timestep)]
-
-        X = self.table['file_path'].tolist()
-        X_lens = self.table['length'].tolist()
-        self.num_samples = len(X)
-        print('[Dataset] - Number of individual training instances:', self.num_samples)
-
-        # Use bucketing to allow different batch size at run time
-        self.X = []
-        batch_x, batch_len = [], []
-
-        for x, x_len in zip(X, X_lens):
-            batch_x.append(x)
-            batch_len.append(x_len)
-            
-            # Fill in batch_x until batch is full
-            if len(batch_x) == bucket_size:
-                # Half the batch size if seq too long
-                if (bucket_size >= 2) and (max(batch_len) > HALF_BATCHSIZE_TIME) and self.sample_length == 0:
-                    self.X.append(batch_x[:bucket_size//2])
-                    self.X.append(batch_x[bucket_size//2:])
-                else:
-                    self.X.append(batch_x)
-                batch_x, batch_len = [], []
-        
-        # Gather the last batch
-        if len(batch_x) > 1: 
-            self.X.append(batch_x)
-
-    def _sample(self, x):
-        if self.sample_length <= 0: return x
-        if len(x) < self.sample_length: return x
-        idx = random.randint(0, len(x)-self.sample_length)
-        return x[idx:idx+self.sample_length]
-
-    def __len__(self):
-        return len(self.X)
-
-    def collate_fn(self, items):
-        items = items[0] # hack bucketing
-        return items
-
-
-class KaldiAcousticDataset(AcousticDataset):
+class KaldiAcousticDataset(FeatDataset):
     
     def __init__(self, extracter, task_config, bucket_size, file_path, sets, 
                  max_timestep=0, libri_root=None, **kwargs):
@@ -119,7 +46,7 @@ class KaldiAcousticDataset(AcousticDataset):
         return generate_masked_acoustic_model_data(spec=(x_pad_batch,), config=self.task_config)
 
 
-class OnlineAcousticDataset(AcousticDataset):
+class OnlineAcousticDataset(FeatDataset):
     
     def __init__(self, extracter, task_config, bucket_size, file_path, sets, 
                  max_timestep=0, libri_root=None, target_level=-25, **kwargs):
